--- sources/presentation/Stride.Core.Presentation.Wpf/Dirtiables/EmptyDirtyingOperation.cs
@@ -1,25 +1,0 @@
-// Copyright (c) .NET Foundation and Contributors (https://dotnetfoundation.org/ & https://stride3d.net) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
-// Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
-using System.Collections.Generic;
-using Stride.Core.Annotations;
-
-namespace Stride.Core.Presentation.Dirtiables
-{
-    public sealed class EmptyDirtyingOperation : DirtyingOperation
-    {
-        public EmptyDirtyingOperation([NotNull] IEnumerable<IDirtiable> dirtiables)
-            : base(dirtiables)
-        {
-        }
-
-        /// <inheritdoc/>
-        protected override void Undo()
-        {
-        }
-
-        /// <inheritdoc/>
-        protected override void Redo()
-        {
-        }
-    }
-}
