+++ sources/presentation/Stride.Core.Presentation/Dirtiables/CollectionChangeOperation.cs
@@ -1,62 +1,60 @@
 // Copyright (c) .NET Foundation and Contributors (https://dotnetfoundation.org/ & https://stride3d.net) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
 // Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
-using System;
-using System.Collections;
-using System.Collections.Generic;
-using System.Collections.Specialized;
-using System.Linq;
-using Stride.Core.Annotations;
-using Stride.Core.Presentation.Dirtiables;
 
-namespace Stride.Core.Presentation.ViewModel
+using System.Collections;
+using System.Collections.Specialized;
+
+namespace Stride.Core.Presentation.Dirtiables
 {
     public class CollectionChangeOperation : DirtyingOperation
     {
         private readonly int index;
-        private IList list;
-        private IReadOnlyCollection<object> items;
+        private IList? list;
+        private IReadOnlyCollection<object>? items;
 
-        private CollectionChangeOperation([NotNull] IList list, NotifyCollectionChangedAction actionToUndo, [NotNull] IEnumerable<IDirtiable> dirtiables)
+        private CollectionChangeOperation(IList list, NotifyCollectionChangedAction actionToUndo, IEnumerable<IDirtiable> dirtiables)
             : base(dirtiables)
         {
-            if (list == null) throw new ArgumentNullException(nameof(list));
+            ArgumentNullException.ThrowIfNull(list);
             if (actionToUndo == NotifyCollectionChangedAction.Reset) throw new ArgumentException("Reset is not supported by the undo stack.");
+
             ActionToUndo = actionToUndo;
             this.list = list;
         }
 
-        public CollectionChangeOperation([NotNull] IList list, NotifyCollectionChangedAction actionToUndo, [NotNull] IReadOnlyCollection<object> items, int index, [NotNull] IEnumerable<IDirtiable> dirtiables)
+        public CollectionChangeOperation(IList list, NotifyCollectionChangedAction actionToUndo, IReadOnlyCollection<object> items, int index, IEnumerable<IDirtiable> dirtiables)
             : this(list, actionToUndo, dirtiables)
         {
-            if (items == null) throw new ArgumentNullException(nameof(items));
+            ArgumentNullException.ThrowIfNull(items);
+
             this.items = items;
             this.index = index;
         }
 
-        public CollectionChangeOperation([NotNull] IList list, [NotNull] NotifyCollectionChangedEventArgs args, [NotNull] IEnumerable<IDirtiable> dirtiables)
+        public CollectionChangeOperation(IList list, NotifyCollectionChangedEventArgs args, IEnumerable<IDirtiable> dirtiables)
             : this(list, args.Action, dirtiables)
         {
             switch (args.Action)
             {
                 case NotifyCollectionChangedAction.Add:
-                    items = args.NewItems.Cast<object>().ToArray();
+                    items = args.NewItems?.Cast<object>().ToArray();
                     index = args.NewStartingIndex;
                     break;
                 case NotifyCollectionChangedAction.Move:
                     // Intentionally ignored, move in collection are not tracked
                     return;
                 case NotifyCollectionChangedAction.Remove:
-                    items = args.OldItems.Cast<object>().ToArray();
+                    items = args.OldItems?.Cast<object>().ToArray();
                     index = args.OldStartingIndex;
                     break;
                 case NotifyCollectionChangedAction.Replace:
-                    items = args.OldItems.Cast<object>().ToArray();
+                    items = args.OldItems?.Cast<object>().ToArray();
                     index = args.OldStartingIndex;
                     break;
                 case NotifyCollectionChangedAction.Reset:
                     throw new NotSupportedException("Reset is not supported by the undo stack.");
                 default:
-                    items = new object[] { };
+                    items = Array.Empty<object>();
                     index = -1;
                     break;
             }
@@ -82,7 +80,9 @@ namespace Stride.Core.Presentation.ViewModel
         /// <inheritdoc/>
         protected override void Undo()
         {
-            int i = 0;
+            if (list is null || items is null) return;
+
+            var i = 0;
             switch (ActionToUndo)
             {
                 case NotifyCollectionChangedAction.Add:
