+++ sources/presentation/Stride.Core.Presentation/Dirtiables/EmptyDirtyingOperation.cs
@@ -1,0 +1,22 @@
+// Copyright (c) .NET Foundation and Contributors (https://dotnetfoundation.org/ & https://stride3d.net) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
+// Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
+
+namespace Stride.Core.Presentation.Dirtiables;
+
+public sealed class EmptyDirtyingOperation : DirtyingOperation
+{
+    public EmptyDirtyingOperation(IEnumerable<IDirtiable> dirtiables)
+        : base(dirtiables)
+    {
+    }
+
+    /// <inheritdoc/>
+    protected override void Undo()
+    {
+    }
+
+    /// <inheritdoc/>
+    protected override void Redo()
+    {
+    }
+}
