+++ sources/presentation/Stride.Core.Presentation/Dirtiables/IDirtyingOperation.cs
@@ -1,0 +1,17 @@
+// Copyright (c) .NET Foundation and Contributors (https://dotnetfoundation.org/ & https://stride3d.net) and Silicon Studio Corp. (https://www.siliconstudio.co.jp)
+// Distributed under the MIT license. See the LICENSE.md file in the project root for more information.
+
+namespace Stride.Core.Presentation.Dirtiables;
+
+public interface IDirtyingOperation
+{
+    /// <summary>
+    /// Gets whether this operation is currently realized.
+    /// </summary>
+    bool IsDone { get; }
+
+    /// <summary>
+    /// Gets the dirtiable objects associated to this operation, or <c>null</c> if no dirtiable is associated.
+    /// </summary>
+    IReadOnlyList<IDirtiable> Dirtiables { get; }
+}
